Issue #1404: AclPermissionApiService moved to acl package
